+++ client/auth/src/components/AppLogout.vue
@@ -16,13 +16,7 @@
 </template>
 
 <script setup>
-import authApi from '@/auth/src/api/auth';
 import Card from 'primevue/card';
-import { onMounted } from 'vue';
-
-onMounted(() => {
-  authApi.logout();
-});
 </script>
 
 <style lang="scss" scoped>
