+++ server/src/routes/tasks/utils.ts
@@ -3,9 +3,7 @@ import { expr } from '@mikro-orm/core';
 const DESC_ORDER_ATTRS = ['priority'];
 const NAME_ATTR = 'name';
 
-export function getNameOrderBy() {
-  return { [expr('lower(name)')]: 'ASC' };
-}
+const getNameOrderBy = () => ({ [expr('lower(name)')]: 'ASC' });
 
 export function processOrderBy(orderBy: string) {
   const isName = orderBy === NAME_ATTR;
